EmulatorJS almost working
The current implementation of EmulatorJS is working using links to the online demo and an explicit link to a NES ROM file.

Next steps are to implement the emu using local files (both emu files and ROM files), and make the system run across multiple systems and games on demand.